+++ apps/prairielearn/src/pages/instructorInstanceAdminSettings/instructorInstanceAdminSettings.html.tsx
@@ -33,7 +33,6 @@ export function InstructorInstanceAdminSettings({
   studentLink,
   publicLink,
   selfEnrollLink,
-  enrollmentManagementEnabled,
   infoCourseInstancePath,
   isDevMode,
 }: {
@@ -51,7 +50,6 @@ export function InstructorInstanceAdminSettings({
   studentLink: string;
   publicLink: string;
   selfEnrollLink: string;
-  enrollmentManagementEnabled: boolean;
   infoCourseInstancePath: string;
   isDevMode: boolean;
 }) {
@@ -66,8 +64,6 @@ export function InstructorInstanceAdminSettings({
     long_name: courseInstance.long_name ?? '',
     display_timezone: courseInstance.display_timezone,
     group_assessments_by: courseInstance.assessments_group_by,
-    show_in_enroll_page:
-      courseInstance.hide_in_enroll_page == null ? true : !courseInstance.hide_in_enroll_page,
     self_enrollment_enabled: courseInstance.self_enrollment_enabled,
     self_enrollment_use_enrollment_code: courseInstance.self_enrollment_use_enrollment_code,
     self_enrollment_restrict_to_institution: courseInstance.self_enrollment_restrict_to_institution,
@@ -230,7 +226,6 @@ export function InstructorInstanceAdminSettings({
               hasModernPublishing={courseInstance.modern_publishing}
               control={control}
               trigger={trigger}
-              enrollmentManagementEnabled={enrollmentManagementEnabled}
               studentLink={studentLink}
               selfEnrollLink={selfEnrollLink}
               enrollmentCode={courseInstance.enrollment_code}