+++ apps/prairielearn/src/pages/instructorInstanceAdminSettings/instructorInstanceAdminSettings.tsx
@@ -26,7 +26,6 @@ import {
   MultiEditor,
   propertyValueWithDefault,
 } from '../../lib/editors.js';
-import { features } from '../../lib/features/index.js';
 import { courseRepoContentUrl } from '../../lib/github.js';
 import { getPaths } from '../../lib/instructorFiles.js';
 import { formatJsonWithPrettier } from '../../lib/prettier.js';
@@ -104,12 +103,6 @@ router.get(
 
     const canEdit = authz_data.has_course_permission_edit && !course.example_course;
 
-    const enrollmentManagementEnabled = await features.enabled('enrollment-management', {
-      institution_id: institution.id,
-      course_id: course.id,
-      course_instance_id: courseInstance.id,
-    });
-
     res.send(
       PageLayout({
         resLocals: res.locals,
@@ -137,7 +130,6 @@ router.get(
                 studentLink={studentLink}
                 publicLink={publicLink}
                 selfEnrollLink={selfEnrollLink}
-                enrollmentManagementEnabled={enrollmentManagementEnabled}
                 infoCourseInstancePath={infoCourseInstancePath}
                 isDevMode={config.devMode}
               />
@@ -162,7 +154,6 @@ router.post(
     const {
       course_instance: courseInstance,
       course,
-      institution,
       urlPrefix,
     } = extractPageContext(res.locals, {
       pageType: 'courseInstance',
@@ -170,12 +161,21 @@ router.post(
     });
 
     if (req.body.__action === 'copy_course_instance') {
-      const { short_name, long_name, start_date, end_date } = z
+      const {
+        short_name,
+        long_name,
+        start_date,
+        end_date,
+        self_enrollment_enabled,
+        self_enrollment_use_enrollment_code,
+      } = z
         .object({
           short_name: z.string().trim(),
           long_name: z.string().trim(),
           start_date: z.string(),
           end_date: z.string(),
+          self_enrollment_enabled: z.boolean(),
+          self_enrollment_use_enrollment_code: z.boolean(),
         })
         .parse(req.body);
 
@@ -234,6 +234,25 @@ router.post(
             }
           : undefined;
 
+      const selfEnrollmentEnabled = propertyValueWithDefault(
+        undefined,
+        self_enrollment_enabled,
+        true,
+      );
+      const selfEnrollmentUseEnrollmentCode = propertyValueWithDefault(
+        undefined,
+        self_enrollment_use_enrollment_code,
+        false,
+      );
+
+      const resolvedSelfEnrollment =
+        (selfEnrollmentEnabled ?? selfEnrollmentUseEnrollmentCode) !== undefined
+          ? {
+              enabled: selfEnrollmentEnabled,
+              useEnrollmentCode: selfEnrollmentUseEnrollmentCode,
+            }
+          : undefined;
+
       // First, use the editor to copy the course instance
       const courseInstancesPath = path.join(course.path, 'courseInstances');
       const editor = new CourseInstanceCopyEditor({
@@ -243,6 +262,7 @@ router.post(
         course_instance: updatedCourseInstance,
         metadataOverrides: {
           publishing: resolvedPublishing,
+          selfEnrollment: resolvedSelfEnrollment,
         },
       });
 
@@ -313,12 +333,6 @@ router.post(
         parsedBody.group_assessments_by,
         'Set',
       );
-      courseInstanceInfo.hideInEnrollPage = propertyValueWithDefault(
-        courseInstanceInfo.hideInEnrollPage,
-        !parsedBody.show_in_enroll_page,
-        false,
-      );
-
       // dates from 'datetime-local' inputs are in the format 'YYYY-MM-DDTHH:MM', and we need them to include seconds.
       const parseDateTime = (date: string) => {
         if (date === '') return undefined;
@@ -329,7 +343,6 @@ router.post(
         courseInstanceInfo.selfEnrollment?.enabled,
         parsedBody.self_enrollment_enabled,
         true,
-        { isUIBoolean: true },
       );
       const selfEnrollmentUseEnrollmentCode = propertyValueWithDefault(
         courseInstanceInfo.selfEnrollment?.useEnrollmentCode,
@@ -359,20 +372,9 @@ router.post(
           selfEnrollmentRestrictToInstitution ??
           selfEnrollmentBeforeDate) !== undefined;
 
-      const enrollmentManagementEnabled = await features.enabled('enrollment-management', {
-        institution_id: institution.id,
-        course_id: course.id,
-        course_instance_id: courseInstance.id,
-      });
       // Only write self enrollment settings if they are not the default values.
       // When JSON.stringify is used, undefined values are not included in the JSON object.
       if (hasSelfEnrollmentSettings) {
-        if (!enrollmentManagementEnabled) {
-          throw new error.HttpStatusError(
-            400,
-            'Self enrollment settings cannot be changed when enrollment management is not enabled.',
-          );
-        }
         if (!courseInstance.modern_publishing) {
           throw new error.HttpStatusError(
             400,